requirements.txt
PrettyTable
mysql-connector-python